requirements.txt
asgiref==3.2.3
dj-database-url==0.5.0
Django==3.0.3
django-configurations==2.2
factory-boy==2.12.0
Faker==4.0.0
python-dateutil==2.8.1
python-decouple==3.3
pytz==2019.3
six==1.14.0
sqlparse==0.3.0
text-unidecode==1.3
